add test kpkd

+++ pndbotics_sdk_py/idl/adam_u/msg/dds_/_LowState_.py
@@ -23,7 +23,5 @@ import cyclonedds.idl.types as types
 @annotate.autoid("sequential")
 class LowState_(idl.IdlStruct, typename="adam_u.msg.dds_.LowState_"):
     motor_state: types.sequence['pndbotics_sdk_py.idl.adam_u.msg.dds_.MotorState_']
-    wireless_remote: types.array[types.uint8, 40]
+    wireless_remote: types.array[types.int16, 40]
     reserve: types.uint32
-
-
